refactor: share otp flow helpers
