+++ src/Chat.Web/wwwroot/js/site.js
@@ -1,4 +1,17 @@
-﻿// Lightweight frontend logger with trace correlation (jQuery-free)
+﻿import {
+    clearOtpTimers,
+    ensureOtpFlow,
+    parseJsonWhenOk,
+    postJson,
+    resetOtpIndicator,
+    scheduleOtpCountdown,
+    scheduleRetryCooldown,
+    setOtpError,
+    showOtpErrorIndicator,
+    showOtpSuccessIndicator
+} from './otpFlow.js';
+
+// Lightweight frontend logger with trace correlation (jQuery-free)
 (function(){
     if (globalThis.appLogger) return; // singleton
     let currentTraceId = null;
@@ -64,113 +77,6 @@
         });
 })();
 
-function postJson(url, data) {
-    return fetch(url, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        credentials: 'same-origin',
-        body: JSON.stringify(data || {})
-    });
-}
-
-function setOtpError(msg) {
-    const el = document.getElementById('otpError');
-    if (!el) return;
-    if (msg) {
-        el.textContent = msg;
-        el.classList.remove('d-none');
-    } else {
-        el.textContent = '';
-        el.classList.add('d-none');
-    }
-}
-
-function ensureOtpFlow() {
-    globalThis.__otpFlow = globalThis.__otpFlow || {};
-    return globalThis.__otpFlow;
-}
-
-function resetOtpIndicator(indicator, retryLink, retryCountdown) {
-    if (indicator) {
-        indicator.classList.remove('d-none', 'fade-hidden');
-        indicator.querySelectorAll('[data-state]').forEach(node => node.classList.add('d-none'));
-        indicator.querySelector('[data-state="sending"]')?.classList.remove('d-none');
-    }
-
-    if (retryLink) {
-        retryLink.classList.add('disabled-link');
-        retryLink.dataset.cooldownLeft = '0';
-    }
-
-    if (retryCountdown) {
-        retryCountdown.textContent = '';
-    }
-}
-
-function scheduleOtpCountdown(flow, countdownEl, countdownTotalMs) {
-    let remainingMs = countdownTotalMs;
-    if (countdownEl) {
-        countdownEl.textContent = Math.ceil(remainingMs / 1000);
-    }
-
-    if (flow.countdownInterval) {
-        clearInterval(flow.countdownInterval);
-    }
-
-    flow.countdownInterval = setInterval(() => {
-        remainingMs -= 1000;
-        if (remainingMs <= 0) {
-            if (countdownEl) {
-                countdownEl.textContent = '0';
-            }
-
-            clearInterval(flow.countdownInterval);
-            flow.countdownInterval = null;
-            return;
-        }
-
-        if (countdownEl) {
-            countdownEl.textContent = Math.ceil(remainingMs / 1000);
-        }
-    }, 1000);
-}
-
-function scheduleRetryCooldown(flow, retryLink, retryCountdown, retryCooldownMs) {
-    if (!retryLink) {
-        return;
-    }
-
-    let left = retryCooldownMs;
-    retryLink.classList.add('disabled-link');
-    retryLink.dataset.cooldownLeft = left.toString();
-    if (retryCountdown) {
-        retryCountdown.textContent = '(' + Math.ceil(left / 1000) + 's)';
-    }
-
-    if (flow.retryInterval) {
-        clearInterval(flow.retryInterval);
-    }
-
-    flow.retryInterval = setInterval(() => {
-        left -= 1000;
-        if (left <= 0) {
-            clearInterval(flow.retryInterval);
-            flow.retryInterval = null;
-            retryLink.classList.remove('disabled-link');
-            retryLink.dataset.cooldownLeft = '0';
-            if (retryCountdown) {
-                retryCountdown.textContent = '';
-            }
-            return;
-        }
-
-        retryLink.dataset.cooldownLeft = left.toString();
-        if (retryCountdown) {
-            retryCountdown.textContent = '(' + Math.ceil(left / 1000) + 's)';
-        }
-    }, 1000);
-}
-
 function scheduleResendAvailability(flow, resendButton, resendDelayMs) {
     if (!resendButton) {
         return;
@@ -199,42 +105,18 @@ function scheduleResendAvailability(flow, resendButton, resendDelayMs) {
     flow.resendAvailInterval = setInterval(updateLabel, 1000);
 }
 
-function clearOtpTimers(flow) {
-    if (flow.startTimeout) {
-        clearTimeout(flow.startTimeout);
-        flow.startTimeout = null;
-    }
-
-    if (flow.countdownInterval) {
-        clearInterval(flow.countdownInterval);
-        flow.countdownInterval = null;
-    }
-}
-
-function showOtpSendSuccess(indicator) {
-    if (!indicator) {
-        return;
-    }
-
-    indicator.querySelectorAll('[data-state]').forEach(node => node.classList.add('d-none'));
-    indicator.querySelector('[data-state="success"]')?.classList.remove('d-none');
-    setTimeout(() => { indicator.classList.add('fade-hidden'); }, 2000);
-    setTimeout(() => { indicator.classList.add('d-none'); }, 2500);
-}
-
 function handleOtpSendFailure(flow, indicator, retryLink, retryCountdown, retryCooldownMs, aborted) {
     if (!indicator) {
         return;
     }
 
-    indicator.querySelectorAll('[data-state]').forEach(node => node.classList.add('d-none'));
     if (aborted) {
         indicator.classList.add('fade-hidden');
         setTimeout(() => indicator.classList.add('d-none'), 250);
         return;
     }
 
-    indicator.querySelector('[data-state="error"]')?.classList.remove('d-none');
+    showOtpErrorIndicator(indicator);
     scheduleRetryCooldown(flow, retryLink, retryCountdown, retryCooldownMs);
 }
 
@@ -336,13 +218,7 @@ document.addEventListener('DOMContentLoaded', () => {
         flow.verifyInFlight = true;
         const verifyStart = performance.now();
         postJson('/api/auth/verify', { userName, code })
-            .then(r => {
-                if (!r.ok) {
-                    throw new Error(globalThis.i18n.InvalidVerificationCode || 'Invalid code');
-                }
-
-                return r.json().catch(() => ({}));
-            })
+            .then(r => parseJsonWhenOk(r, globalThis.i18n.InvalidVerificationCode || 'Invalid code'))
             .then(() => {
                 appLogger.info('OTP verify success', { user: userName });
                 if (flow.lastSendId) {
@@ -449,20 +325,14 @@ document.addEventListener('DOMContentLoaded', () => {
             }
         }, timeoutMs);
 
-        fetch('/api/auth/start', { method:'POST', headers:{'Content-Type':'application/json'}, body: JSON.stringify({ userName }), credentials:'same-origin', signal: controller.signal })
-            .then(r => {
-                if (!r.ok) {
-                    throw new Error(globalThis.i18n.FailedToSendCode || 'Failed to send code');
-                }
-
-                return r.json().catch(() => ({}));
-            })
+        postJson('/api/auth/start', { userName }, { signal: controller.signal })
+            .then(r => parseJsonWhenOk(r, globalThis.i18n.FailedToSendCode || 'Failed to send code'))
             .then(() => {
                 if (flow.activeUser !== userName || controller.signal.aborted) return;
                 flow.lastSendCompletedTs = performance.now();
                 const durMs = Math.round(flow.lastSendCompletedTs - flow.lastSendStartTs);
                 appLogger.info('OTP code sent', { user: userName, durationMs: durMs, sendId });
-                showOtpSendSuccess(indicator);
+                showOtpSuccessIndicator(indicator);
                 // Move to step2 only on initial send, keep on step2 for resend
                 if (!isResend){
                     const step1 = $('#otp-step1');
@@ -580,8 +450,7 @@ document.addEventListener('DOMContentLoaded', () => {
             }
             const resendBtn = document.getElementById('btn-resend-otp');
             if (resendBtn) resendBtn.disabled = false;
-            clearOtpTimers(flow);
-            if (flow.retryInterval) { clearInterval(flow.retryInterval); flow.retryInterval=null; }
+            clearOtpTimers(flow, { retryInterval: true, resendAvailInterval: true });
             const countdownEl = document.getElementById('otpSendCountdown'); if (countdownEl) countdownEl.textContent='0';
             const retryCountdown = document.getElementById('otpRetryCountdown'); if (retryCountdown) retryCountdown.textContent='';
             const retryLink = document.getElementById('otpRetryLink'); if (retryLink) { retryLink.classList.add('disabled-link'); retryLink.dataset.cooldownLeft='0'; }
